Added While Loop to the Run_Game function
Created a While loop that tells the user / player how many attempts they have remaining before losing.

Created the While loop to keep asking for the user's guess as long as they have more than 0 attempts.

Also added feedback to user if they enter an incorrect key.

Added feedback to inform user whether or not they guessed correctly and if the random word has that letter in it.